refactor: :recycle: Extracting QuestionFlow methods into a concern
